% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vs}
\alias{vs}
\title{Polling station data}
\format{
A \code{tibble} with 1,384 rows and 3 columns
}
\source{
Collected from \url{https://valgdatabase.dst.dk/?lang=en}
}
\usage{
vs
}
\description{
Polling station data with IDs
}
\keyword{datasets}
