% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPS.R
\name{getPS}
\alias{getPS}
\title{Get voting station data}
\usage{
getPS(spatial = TRUE)
}
\arguments{
\item{spatial}{a logical value. Return data of class \code{sf}? Defaults to \code{TRUE}}
}
\value{
data on danish voting stations
}
\description{
This function extracts spatial data on voting stations in Denmark from \url{https://info.aws.dk/afstemningsomraader/}
}
\examples{
voting_stations <- getPS()
}
\author{
Jeppe Vierø
}
